feat(resultsAppSetupFE): addressed PR reviews comments, added ENUM for storing Views
