+++ src/Analysis/GWASResultsV2/SharedComponents/ReturnHomeButton/ReturnHomeButton.jsx
@@ -2,6 +2,7 @@ import React, { useContext } from 'react';
 import { Button } from 'antd';
 import { ArrowLeftOutlined } from '@ant-design/icons';
 import SharedContext from '../../Utils/SharedContext';
+import VIEWS from '../../Utils/ViewsEnumeration';
 
 const ReturnHomeButton = () => {
   const { setCurrentView } = useContext(SharedContext);
@@ -9,7 +10,7 @@ const ReturnHomeButton = () => {
     <Button
       type='text'
       icon={<ArrowLeftOutlined />}
-      onClick={() => setCurrentView('home')}
+      onClick={() => setCurrentView(VIEWS.home)}
     >
       Return Home Button
     </Button>
