+++ src/Analysis/GWASResultsV2/Utils/ViewsEnumeration.js
@@ -1,0 +1,7 @@
+const VIEWS = {
+  home: 'home',
+  results: 'results',
+  execution: 'execution',
+};
+
+export default VIEWS;
