+++ src/Analysis/GWASResultsV2/Views/Home/Utils/GetTableDataFromApi.js
